ui (vote) Move candidate search under coalition index
When search box was on top, it was unclear what was supposed
to happen when writing a search term, when only the the coalition
index was visible on the screen. Now the search box is next to
the candidate list, which allows the user to see the filtering
while he is typing.